requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9 # PostgreSQL adapter
pika==1.3.2
redis==5.0.1
tenacity==8.2.3 # For retries
python-jose[cryptography]==3.3.0 # For JWT
alembic==1.12.0 # Add Alembic dependency
